Application/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==3.2
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
nltk==3.7
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.13
